🔨 Add useInterval Hook, Refactor Live to use useInterval
Added a useInterval Hook (demo here: https://codesandbox.io/s/sessiontimer-useinterval-hook-demo-0l6jq)

Refactored Live workspace item and it's dependents to use hooks.
