+++ packages/app/src/app/hooks/index.ts
@@ -1,1 +1,2 @@
+export { useInterval } from './useInterval';
 export { useScript } from './useScript';
