+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/items/Live/LiveButton/index.ts
@@ -1,0 +1,1 @@
+export { LiveButton } from './LiveButton';
