+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/items/Live/LiveInfo/index.ts
@@ -1,0 +1,1 @@
+export { LiveInfo } from './LiveInfo';
