+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/items/Live/index.ts
@@ -1,0 +1,1 @@
+export { Live } from './Live';
